Initial project set up.
When starting a new project, I like to:

1. Provide a README explaining it's high-level use, as well as how to
get started developing and running the application
2. If open source, I include a contribor agreement, license,
contributing guidelines, etc.
3. Set up the test structure
4. Set up the web app structure

Technically, this commit can be deployed to Heroku *right now*. It would
just be very very boring.

Because of this I can make deploying and verifying a deployed version
part of my workflow