Automate loading of sample data into database
When working on a project it is useful to have sample data. We've
provided a `rake` task to populate our database with 10 gibberish
overheards.

New Libraries
=============

`rake` - This is a tool for running one-off ruby commands without
writing a ton of small scripts. You define `tasks`, then
run them with `rake <task-name>`. To get a list of available tasks, run
`rake -T -A`.

New Concepts
============

`Seeding` - We use the phrase "seeding" to mean "populating with sample
data"