Ensure HTML is Valid HTML5
Invalid HTML is the leading cause of rendering errors, cross-browser
frustration, and depression in front end developers.

To validate I:
1. Right clicked the page in my browser and clicked "view source"
2. Copied the source
3. Visited `http://html5.validator.nu`
4. Selected 'Text Field' from
5. Pasted in my page source
6. Clicked Validate
7. Fixed the errors it told me about
8. Goto 1

New Tags
========
`<!DOCTYPE html>` - Tells the browser that this is an HTML5 document.
[Reference](https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/HTML5/Introduction_to_HTML5#Declaring_that_the_document_contains_HTML5_mark-up_with_the_HTML5_doctype)

`<head>` - Contains elements which provide
[metadata](https://github.com/codeunion/fundamentals-of-web-development/wiki/Glossary#metadata)
about the page.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/head)

`<title>` - Declares the title to be shown in the browsers title bar.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/title)

`<body>` - Contains the content of the web page.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/body)

`<header>` - Contains introduction and/or navigation.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/header)

`<main>` - Contains the main content of the page.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/main)

`<h1>` - The most important heading on the page.
[Reference](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements)