Test creating an overheard with a form
Again we see the pattern of imaging the user interface, then codifying
it into a test. To get started, I
[copy-pastad](http://imgbit.com/images/652a2231491269586521.jpg) the
test/features/list_overheards_test.rb and renamed the test class and
test method.

New Methods
===========

`click_on` - Find and clicks a button or link.  Provided by
`Capybara::DSL`.
[Reference](http://rubydoc.info/github/jnicklas/capybara/master/Capybara/Node/Actions:click_link_or_button)

`fill_in` - Find a field and fill in a form field. Field must be labeled
with the text of the first argument.  Fills in the field with the
`:with` keys value from the options hash.  Provided by `Capybara::DSL`.
[Reference](http://rubydoc.info/github/jnicklas/capybara/master/Capybara/Node/Actions#fill_in-instance_method)