Create overheard in database on form submission
BEHOLD! We now are storing overheards in the database! These 4 lines are
the heart and soul of our program:

1. We define a
[route](https://github.com/codeunion/fundamentals-of-web-development/wiki/Glossary#route-sinatra)
that builds responses for POST requests to /overheards.
2. When a request comes in, we get the request data from
params["overheard"]. If we `p params` It will look something like `{
  "body" => "Whatever the user wrote" }`
3. We pass that data to the `Overheard.create` method that
`Datamapper::Model` so generously provides.
4. We send a response to the browser to `redirect` it to the "/" route.

New Methods
===========

`post` - Define a route that builds responses to POST requests.

`redirect` - Sends a response to the requester with the [301 status
code](http://en.wikipedia.org/wiki/HTTP_301) and a location of the
passed in string. Causes a whole-nother [request response
cycle](https://github.com/codeunion/fundamentals-of-web-development/wiki/Glossary#requestresponse-cycle)

`Datamapper::Model#create` - Creates a row in the models table with the
attributes passed in. Returns an instance of the model.
[Reference](http://rdoc.info/github/datamapper/dm-core/DataMapper/Model#create-instance_method)