Extract default layout
Because the two views were pretty much the exact same, it's a perfect
opportunity to extract a layout. Now we only have to change
`views/layout.erb` if we make any layout changes.

The secret sauce of a layout is calling `yield` within the template.
This is how you tell erb to render the template specified within the
route.

New Syntax
==========
`<%= foo %>` - This `erb` syntax says to evaluate the code between `<%=`
and `%>` as ruby and print out the result.