Display overheard body error to the user
Bam! We've got errors being shown to the user. Important things to
consider:

1. `views` have access to `@instance_variables` defined in routes. This
   means we can access the `@overheard` variable inside of our
   `new_overheard.erb` view.
2. Every `Datamapper::Model` has an `errors` method that contains
   validation errors that were caught when trying to create or validate
   the model.
3. I had to update the assertion that tested for the error because I
   used the wrong words. Even though the page was showing the error, it
   didn't exactly match what I expected.

New Methods
===========

`Datamapper::Model#errors` - Provides a hash of errors caught when
validating a model. Each property of the model is a key in the errors
hash, and an array of validation failures is the value. If there are no
validation errors for an attribute, it returns an empty array.