Provider errors on invalid API calls
When the overheard added via the api can't be saved; we want to let the
user know that there was an error and on which fields.

The Datamapper::Resource `errors` object is a kind of Enumerable; so
that means we can call `.to_h` to get the errors as a hash where the
attributes are the keys and an array of errors are the value. This is
perfect for our api consumers.

Unfortunately we're starting to get a bit of duplication that is kind of
nasty. We'll want to fix that soon...