Overheards may be listed as JSON
A relatively tiny change to the `get "/"` route allows us to expose a
json representation of our overheards! Hooray!

It also looks like we're getting some duplication; especially with our
`if request.accept?("whatever")` and `JSON.dump` bits.

This is an indication we could do some refactoring!

Why I changed the tests:
------------------------
* Again I misused `rack-test`. If you want to send headers, apparently
  you have to give it a body, even if it's a GET request! Ah well.
* `assert_equal` provides better output on failure than just `assert`.